move mock npm to testing file
