+++ analyzers/nodejs/nodejs_test.go
@@ -1,44 +1,17 @@
 package nodejs_test
 
 import (
-	"os"
 	"testing"
-
-	"github.com/fossas/fossa-cli/files"
 
 	"github.com/stretchr/testify/assert"
 
 	"github.com/fossas/fossa-cli/analyzers/nodejs"
-	"github.com/fossas/fossa-cli/buildtools/npm"
 	"github.com/fossas/fossa-cli/module"
 	"github.com/fossas/fossa-cli/pkg"
 )
 
 func TestNoDependencies(t *testing.T) {
 	t.Skip("unimplemented")
-}
-
-type MockNPM struct{}
-
-func (n MockNPM) List(_ string) (npm.Output, error) {
-	var output npm.Output
-	dir, _ := os.Getwd()
-
-	err := files.ReadJSON(&output, dir, "fixtures/npmLsOutput.json")
-
-	if err != nil {
-		panic(err)
-	}
-
-	return output, nil
-}
-
-func (n MockNPM) Clean(dir string) error {
-	return nil
-}
-
-func (n MockNPM) Install(dir string) error {
-	return nil
 }
 
 func TestAnalyzeWithNpmLs(t *testing.T) {
@@ -54,7 +27,7 @@ func TestAnalyzeWithNpmLs(t *testing.T) {
 	analyzer, err := nodejs.New(nodeModule)
 	assert.NoError(t, err)
 
-	analyzer.Tool = MockNPM{}
+	analyzer.Tool = nodejs.MockNPM{}
 
 	analysisResults, err := analyzer.Analyze()
 	assert.NoError(t, err)
